- Updated readme. - Light refactoring. - Accessibility updates.
