+++ app/src/main/java/com/example/rightpoint/omdb/search/models/Search.kt
@@ -1,5 +1,6 @@
-package com.example.rightpoint.omdb.shows
+package com.example.rightpoint.omdb.search.models
 
+import com.example.rightpoint.omdb.shows.Show
 import java.io.Serializable
 
 /**
