+++ app/src/main/res/layout/search_results_view.xml
@@ -2,11 +2,13 @@
 <merge xmlns:android="http://schemas.android.com/apk/res/android">
 
     <Button
-        android:id="@+id/shuffleLayout"
+        android:id="@+id/shuffleAction"
         android:layout_width="48dp"
         android:layout_height="wrap_content"
         android:background="@drawable/ic_shuffle_black_24dp"
         android:layout_gravity="bottom|end"
+        android:focusableInTouchMode="false"
+        android:importantForAccessibility="no"
         />
